Split component-specific ALumni parts into mixins
